+++ Lab2/MyThreadPool/MyThreadPool/MyThreadPool.csproj
@@ -44,12 +44,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MyTask\IMyTask.cs" />
-    <Compile Include="MyTask\MyTask.cs" />
     <Compile Include="MyThreadPool\MyThreadPool.cs" />
     <Compile Include="Exceptions\MyThreadPoolNotWorkingException.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Exceptions\SpecialTestException.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
