+++ Lab2/MyThreadPool/MyThreadPoolTests/MyThreadPoolTests.csproj
@@ -58,6 +58,7 @@
   <ItemGroup>
     <Compile Include="MyThreadPoolTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SpecialTestException.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
