+++ Lab2/MyThreadPool/MyThreadPoolTests/SpecialTestException.cs
